+++ src/main/java/fuzs/goldenagecombat/mixin/accessor/ItemAccessor.java
@@ -2,6 +2,7 @@ package fuzs.goldenagecombat.mixin.accessor;
 
 import net.minecraft.world.item.Item;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
 import java.util.UUID;
@@ -12,4 +13,8 @@ public interface ItemAccessor {
     static UUID getBaseAttackDamageUUID() {
         throw new IllegalStateException();
     }
+
+    @Accessor
+    @Mutable
+    void setMaxStackSize(int maxStackSize);
 }